chore: move mixin to utils package
