+++ maidr/core/plot/bar_plot.py
@@ -6,9 +6,9 @@ from matplotlib.container import BarContainer
 from maidr.core.enum.maidr_key import MaidrKey
 from maidr.core.enum.plot_type import PlotType
 from maidr.core.maidr_plot import MaidrPlot
-from maidr.core.mixin.merger_mixin import DictMergerMixin
-from maidr.core.mixin.extractor_mixin import (
+from maidr.utils.mixin import (
     ContainerExtractorMixin,
+    DictMergerMixin,
     LevelExtractorMixin,
 )
 from maidr.exception.extraction_error import ExtractionError
