+++ maidr/core/plot/box_plot.py
@@ -5,11 +5,11 @@ from matplotlib.axes import Axes
 from maidr.core.enum.maidr_key import MaidrKey
 from maidr.core.enum.plot_type import PlotType
 from maidr.core.maidr_plot import MaidrPlot
-from maidr.core.mixin.extractor_mixin import (
+from maidr.utils.mixin import (
     ContainerExtractorMixin,
+    DictMergerMixin,
     LevelExtractorMixin,
 )
-from maidr.core.mixin.merger_mixin import DictMergerMixin
 from maidr.exception.extraction_error import ExtractionError
