+++ maidr/core/plot/heat_plot.py
@@ -7,11 +7,11 @@ from matplotlib.collections import QuadMesh
 from maidr.core.enum.maidr_key import MaidrKey
 from maidr.core.enum.plot_type import PlotType
 from maidr.core.maidr_plot import MaidrPlot
-from maidr.core.mixin.extractor_mixin import (
+from maidr.utils.mixin import (
+    DictMergerMixin,
     LevelExtractorMixin,
     ScalarMappableExtractorMixin,
 )
-from maidr.core.mixin.merger_mixin import DictMergerMixin
 from maidr.exception.extraction_error import ExtractionError
